backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
python-dotenv
groq
psycopg2-binary
sqlalchemy